Move PageCodecMarker to SPI
