+++ presto-main/src/test/java/com/facebook/presto/execution/buffer/TestThriftBufferResult.java
@@ -13,6 +13,7 @@
  */
 package com.facebook.presto.execution.buffer;
 
+import com.facebook.presto.spi.page.PageCodecMarker;
 import com.google.common.collect.ImmutableList;
 import org.testng.annotations.Test;
