+++ presto-spi/src/main/java/com/facebook/presto/spi/page/PageCodecMarker.java
@@ -11,12 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.execution.buffer;
+package com.facebook.presto.spi.page;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
 
-import static com.google.common.base.Preconditions.checkArgument;
+import static java.lang.String.format;
 
 /**
  * Encodes boolean properties for {@link SerializedPage} by using a bitmasking strategy, allowing
@@ -67,5 +67,12 @@ public enum PageCodecMarker
                 .filter(marker -> marker.isSet(markers))
                 .map(PageCodecMarker::name)
                 .collect(Collectors.joining(", "));
+    }
+
+    private static void checkArgument(boolean condition, String message, Object... messageArgs)
+    {
+        if (!condition) {
+            throw new IllegalArgumentException(format(message, messageArgs));
+        }
     }
 }
